requirements.txt
meetup-api
requests
discord-webhook
pyyaml
pytz
configargparse